refactor(uffs-mft): introduce Usn(i64) newtype for journal USNs
Replaces ad-hoc `i64` USN values with a typed `uffs_mft::usn::Usn`
newtype so cache wrap detection, checkpoint persistence, and journal
record fields stop relying on raw integers that any caller could
mis-interpret as a generic counter.

The newtype is `Copy + Eq + Ord + Hash + Display` and exposes
`Usn::ZERO`, `Usn::new`, `Usn::raw`, and `Usn::is_zero` so:

* monotonic comparisons (`cached_usn < first_usn`, `start_usn >=
  next_usn`) read directly from the type rather than re-deriving
  meaning from `i64` arithmetic;
* the `0` sentinel meaning "no prior checkpoint — read from journal
  head" gets a self-documenting constant;
* tracing fields use `%`-Display so logs render `123` not
  `Usn(123)`.

## Wire format unchanged

Pinned by `tests_extensions::serialize_roundtrip` /
`tests_storage::empty_serialized_index` /
`usn::tests::raw_roundtrip_preserves_i64_exactly`:

* on-disk `IndexHeader.next_usn` still serializes as 8 LE i64 bytes
  via `.raw().to_le_bytes()` and deserializes via `Usn::new(read_i64)`;
* daemon-side `JournalSource::poll` keeps its `cursor: u64`
  persistence contract — the `Usn` wrap/unwrap is a single
  call-boundary narrowing inside `sources.rs`;
* the `#[repr(C)]` Win32 ABI mirror structs (`UsnJournalDataV0`,
  `ReadUsnJournalDataV0`, `UsnRecordV2Header`) keep their raw `i64`
  fields — they mirror `winioctl.h` byte-for-byte;
* the `uffs-mft usn-read --start-usn <i64>` CLI flag stays
  `Option<i64>` and is wrapped at the consumer site.

## File split

Migrating UI doc + tests pushed `usn.rs` over the 800-LOC policy
limit.  Split into `usn/mod.rs` (DTOs + reason flags + aggregation +
non-Windows stubs + tests, 428 LOC) and `usn/windows.rs` (Win32 FFI
mirror structs + ioctl helpers + targeted FRS reads, 527 LOC).  No
file needs a size-policy exception.

## Untouched on purpose

`StandardInfo.usn: u64` (from the NTFS `$STANDARD_INFORMATION`
attribute) is a separate USN flavour with a different signedness
contract; out of scope for this PR.
